# openapi.yaml
openapi: 3.0.0
info:
  title: CivicConnect API
  description: API specification generated from implemented Go backend routes.
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register citizen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required: [email, password]
      responses:
        '201':
          description: Registered
  /api/v1/auth/register-admin:
    post:
      summary: Register admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                adminSecret:
                  type: string
              required: [email, password, adminSecret]
      responses:
        '201':
          description: Registered
        '403':
          description: Invalid secret / disabled
  /api/v1/auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required: [email, password]
      responses:
        '200':
          description: Logged in
  /api/v1/me:
    get:
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
    patch:
      summary: Update current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                oldPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        '200':
          description: Updated user
    delete:
      summary: Delete current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
              required: [password]
      responses:
        '200':
          description: Deleted
  /api/v1/issues:
    get:
      summary: List public issues
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radiusMeters
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
        - name: offset
          in: query
          schema:
            type: number
        - name: status
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Issue list
  /api/v1/issues/stats:
    get:
      summary: Public issue statistics
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radiusMeters
          in: query
          schema:
            type: number
        - name: status
          in: query
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Stats
...
  /api/v1/citizen/issues:
    get:
      summary: Citizen nearby issues
      security:
        - bearerAuth: []
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radiusMeters
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        '200':
          description: Citizen issue list
  /api/v1/citizen/issues/mine:
    get:
      summary: List issues reported by current citizen
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Reporter-specific issues list
    post:
      summary: Create or merge citizen issue
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                imageUrls:
                  type: array
                  items:
                    type: string
                departmentId:
                  type: string
                location:
                  type: object
                  properties:
                    lat:
                      type: number
                    lng:
                      type: number
                  required: [lat, lng]
              required: [title, description, imageUrls, departmentId, location]
      responses:
        '201':
          description: Created
        '200':
          description: Merged
...
  /api/v1/uploads/images:
    post:
      summary: Upload image
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                context:
                  type: string
              required: [file]
      responses:
        '201':
          description: Uploaded
  /api/v1/head/issues/pending:
    get:
      summary: List pending issues for department head
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Pending issues
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            title:
                              type: string
                            description:
                              type: string
                            reporterId:
                              type: string
                            reporterName:
                              type: string
                            reporterEmail:
                              type: string
                            status:
                              type: string
                            category:
                              type: string
                            departmentId:
                              type: string
                            imageUrls:
                              type: array
                              items:
                                type: string
                            createdAt:
                              type: string
                              format: date-time
                            updatedAt:
                              type: string
                              format: date-time
  /api/v1/head/issues/escalations:
    get:
      summary: List escalations for department head
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Escalated issues
...
  /api/v1/head/workers:
    get:
      summary: List department workers
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: includeBlocked
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Worker list
    post:
      summary: Register authority worker
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required: [email, password]
      responses:
        '201':
          description: Worker created
...
  /api/v1/head/authorities:
    post:
      summary: Register authority worker
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required: [email, password]
      responses:
        '201':
          description: Worker created
  /api/v1/authority/issues:
    get:
      summary: List authority issues
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Authority issues
...
  /api/v1/admin/departments:
    get:
      summary: List departments
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Department list
    post:
      summary: Create department
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required: [name]
      responses:
        '201':
          description: Department created
...
  /api/v1/admin/authorities:
    get:
      summary: List authority heads
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Authority head list
    post:
      summary: Register authority head
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                departmentId:
                  type: string
              required: [email, password, departmentId]
      responses:
        '201':
          description: Authority head created
  /api/v1/admin/issues/flagged:
    get:
      summary: List flagged issues
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Flagged issues
  /api/v1/admin/issues/escalations:
    get:
      summary: List escalated issues (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Escalated issues
...